Add Fruit Basket event, Fly enemy, constitution system, new effect types
Excel data:
- convert-excel.js: use parseFloat for enemy weight (was parseInt, truncated 0.5→0)
- Fly enemy: weight 0.5, from Mewgenics
- Leeching Seed item (tagged 'seed'), Keeper's Sack item

Constitution system:
- gameState.startingMaxHealth captured at run start
- constitution = floor((maxHP - startingMaxHP) / 5); +1 per 5 gained max HP
- constitution stat usable in stat_check events (stat: 'constitution')
- Constitution value shown on event choice screen when > 0

New event effect types:
- heal_percent: heals a % of max HP
- spawn_enemies: queues enemy spawns for next fight, bypassing weight budget
  (stored in gameState.pendingSpawnEnemies, applied in showDiceCombatModal)

Fruit Basket event (fruit_basket):
- Eat (Charisma): crit heal 50%, good heal 20%, bad/crit-bad gain Poison 3/4
- Destroy (Strength): crit get Seed item, good/bad nothing, crit-bad spawn 6-8 Fly
- Examine (Intelligence): crit heal 15+Holy Shield, good heal 5,
  bad Ambushed, crit-bad gain 5 Fear

README: document heal_percent, spawn_enemies, constitution stat

+++ convert-excel.js
@@ -325,7 +325,7 @@ const enemies = enemiesData.map(row => {
 
   // Parse weight - may be N/A for spawn-only enemies
   const rawWeight = row['Weight'];
-  const weight = (rawWeight !== undefined && rawWeight !== 'N/A' && rawWeight !== '') ? parseInt(rawWeight) : null;
+  const weight = (rawWeight !== undefined && rawWeight !== 'N/A' && rawWeight !== '') ? parseFloat(rawWeight) : null;
 
   // Parse difficulty - may be N/A for spawn-only enemies
   const difficulty = (row['Difficulty'] && row['Difficulty'] !== 'N/A') ? row['Difficulty'] : null;
